Move migration tests to the correct place
